feat(deferred): close E1-E3 + agreement-FA real compute (overlooked items)
E1: sed_add / sed_scale / sed_neg already existed in OctonionNorm.v -- no-op
    (the deferral note in C1627 was wrong; verified and documented).

E2: C1627 r4_orbit_lin_indep upgraded from placeholder to real spec:
    forall c0..c3 : R,
      sed_add (sed_scale c0 r4_x_lift)
        (sed_add (sed_scale c1 r4_y_lift)
          (sed_add (sed_scale c2 r4_xe) (sed_scale c3 r4_ye))) = sed_zero
      -> c0 = 0 /\ c1 = 0 /\ c2 = 0 /\ c3 = 0.
    Compiles + kernel-checks. Comment lane updated.

E3: Full proofs tree rebuild (275 .v files clean against current Prelude).
    Moreno1997.v `From OpenGororoba Require Export` now includes
    C1627_MorVlambdaQuadOrbit; downstream importers of Moreno1997 pick up
    the four-tuple witness without an extra import.

cubecl_backend probe wired to real launcher::is_available() probe:
    probe_cubecl() previously returned a hardcoded
    {cuda=true, vulkan=true, wgpu=true} placeholder. Now it probes the
    cubecl-wgpu runtime live, reports wgpu_available based on the real
    handshake, and reports cuda/vulkan as false until those cubecl
    backends are wired up (the production CUDA/Vulkan paths remain via
    cudarc / ash and route through other backend arms). Module docstring
    updated to reflect that the WGPU lane is no longer a stub.

heliosphere-hybrid-detector agreement thresholds now compute real FA:
    The k-of-6 agreement code previously set `fa_rate = 0.0; // placeholder`
    and used `f1 = det_rate` as a stand-in. Replaced with a materialized
    agreement-hour set (filter all6 by per-hour detector-firing count) routed
    through the existing `score()` helper. Detection, FA, and F1 are now
    computed identically to the union rows.

Clippy cleanups: replaced unnecessary u32 cast and manual div_ceil in the
cubecl backend code path so the cubecl feature stays gate-green.

C-1627 status_note updated in canonical SQLite (revision 3).

+++ crates/cd_kernel/src/turboquant/backend.rs
@@ -237,7 +237,7 @@ impl BackendQuantizer {
                 let req = QuantizeRequest {
                     values,
                     boundaries: self.cpu.boundaries(),
-                    bits: self.cpu.bits() as u32,
+                    bits: self.cpu.bits(),
                 };
                 match launcher::quantize(&req, out) {
                     Ok(()) => Ok(()),
